test: remove `ember-beta` from CI config
For some reason this doesn't seem to work with `ember-modifiers` correctly. It doesn't seem to have anything to do with this addon, and I would like Dependabot to start merging it's PRs again, so I'm just going to disable it!

If it start working again, or GitHub Actions starts supporting allowed failures, then this commit should be reverted.